6-size.c

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ long int l;
 long long int o;
 float f;
 
-	printf("Size of a char: %lu byte(s) \n", sizeof(h));
-	printf("Size of an int: %lu byte(s) \n", sizeof(i));
-	printf("Size of a long int: %lu byte(s) \n", sizeof(l));
-	printf("Size of a long long int: %lu byte(s) \n", sizeof(o));
-	printf("Size of a float: %lu byte(s) \n", sizeof(f));
+	printf("Size of a char: %lu byte(s) \n", (unsigned long)sizeof(h));
+	printf("Size of an int: %lu byte(s) \n", (unsigned long)sizeof(i));
+	printf("Size of a long int: %lu byte(s) \n", (unsigned long)sizeof(l));
+	printf("Size of a long long int: %lu byte(s) \n", (unsigned long)sizeof(o));
+	printf("Size of a float: %lu byte(s) \n", (unsigned long)sizeof(f));
 	return (0);
 }
